requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 
tabulate 
requests
oauth2client
termcolor